removed a lot of old code and re-established entry points
