+++ .env.example
@@ -21,5 +21,4 @@ UPHOLD_DONOR_CARD_ID=6654ecb0-6079-4f6c-ba58-791cc890a561
 WALLET_COOLDOWN_HRS=24
 VOTING_COHORTS=control,grant,test,ads,safetynet
 ADS_AVAILABLE_LIST=ZZ
-ADS_URL=https://ads-serve.bravesoftware.com
 PER_MEMBER_CLAIM_LIMIT=3
