+++ .git-hooks/pre-commit
@@ -1,1 +1,1 @@
-npm run security
+yarn run security
