+++ base.yml
@@ -1,31 +1,17 @@
 version: "2.1"
 
 services:
-  ledger:
-    image: bat-ledger
+  eyeshade:
+    image: bat-eyeshade
     build: .
     restart: always
     volumes:
       - ./test/secrets:/etc/kafka/secrets
     environment:
-      - "ADS_AVAILABLE_LIST=US"
-      - "ADS_URL=${ADS_URL-https://ads-serve.bravesoftware.com}"
       - "ANON_RATE_LIMIT_PER_M=60000"
-      - "BAT_BALANCE_SERVER=http://balance-web:3003"
       - "BAT_EYESHADE_SERVER=http://eyeshade-web:3002"
-      - "BAT_GRANT_POSTGRES_URL=postgres://grants:password@grant-postgres/grants?sslmode=disable"
-      - "BAT_GRANT_SERVER=http://grant-web:3333"
-      - "BAT_WALLET_MIGRATION_SERVER=http://wallet-migration-web:3333"
-      - "WALLET_MIGRATION_SERVER=http://wallet-migration-web:3333"
-      - "WALLET_MIGRATION_TOKEN=${WALLET_MIGRATION_TOKEN-98f27f2e-239b-40e9-8f62-bf1a55ea530c}"
-      - "BAT_WALLET_MIGRATION_POSTGRES_URL=${BAT_WALLET_MIGRATION_POSTGRES_URL-postgres://wallet-migration:password@wallet-migration-postgres/wallet-migration}"
-      - "BAT_LEDGER_SERVER=http://ledger-web:3001"
-      - "BAT_MONGODB_URI=mongodb://mongo"
       - "BAT_POSTGRES_URL=postgres://eyeshade:password@eyeshade-postgres/eyeshade"
       - "BAT_RATIOS_URL=https://ratios-staging.mercury.basicattentiontoken.org"
-      - "BAT_REDEEMER_REDIS_URL=redis://redeemer-redis:6379"
-      - "BAT_REDEEMER_SERVER=http://redeemer-web:3333"
-      - "BAT_REWARD_SERVER=http://rewards-dev:3333"
       - "BAT_REDIS_URL=redis://redis:6379"
       - "DEBUG=${DEBUG-*,-babel,-eslint:*,-mongo:*,mongo:queries}"
       - "ENABLE_ADS_PAYOUT=1"
@@ -39,14 +25,8 @@ services:
       - "KAFKA_REQUIRED_ACKS=1"
       - "FEATURE_WALLET=true"
       - "FREEZE_SURVEYORS_AGE_DAYS=${FREEZE_SURVEYORS_AGE_DAYS-7}"
-      - "GRANT_SERVER=http://grant-web:3333"
-      - "LEDGER_URL=${LEDGER_URL-http://ledger-web:3001}"
-      - "MONGODB_URI=${MONGODB_URI-mongodb://mongo/ledger}"
       - "NODE_ENV=${NODE_ENV-development}"
       - "PER_MEMBER_CLAIM_LIMIT=4"
-      - "REDEEMER_REDIS_HOST=${REDEEMER_REDIS_HOST-redeemer-redis}"
-      - "REDEEMER_SERVER=${REDEEMER_SERVER-http://redeemer-web:3333}"
-      - "REWARD_SERVER=http://rewards-dev:3333"
       - "REDIS_URL=${REDIS_URL-redis://redis:6379}"
       - "GRANT_TOKEN=596fde96-cc80-40e5-b81f-a21c81bb101a"
       - "ALLOWED_ADS_TOKENS=78102ca7-454d-4f6c-b74d-4e1e38e7e40b"
@@ -58,9 +38,6 @@ services:
       - "ALLOWED_REFERRALS_TOKENS=a819d462-5fbf-40ed-a941-045de57769ca"
       - "ALLOWED_STATS_TOKENS=f9b58f9a-fb23-49eb-a6e2-9715d46f647b"
       - "BAT_SETTLEMENT_ADDRESS=6654ecb0-6079-4f6c-ba58-791cc890a561"
-      - BALANCE_CACHE_TTL_S
-      - BALANCE_TOKEN
-      - BALANCE_URL
       - "BAT_ADS_PAYOUT_ADDRESS=fb599883-c30c-4428-91be-ba901e69a0b2"
       - BAT_FEE_ACCOUNT
       - BAT_RATIOS_TOKEN
@@ -87,10 +64,6 @@ services:
       - GITHUB_ORG
       - HOST
       - IP_GRAYLIST
-      - LEDGER_GITHUB_CLIENT_ID
-      - LEDGER_GITHUB_CLIENT_SECRET
-      - LEDGER_SLACK_CHANNEL
-      - LEDGER_SLACK_ICON_URL
       - MIGRATE_DATABASE_URL
       - MIGRATE_MONGODB_URI
       - PORT
