+++ bat-utils/boot-hapi.js
@@ -1,12 +1,10 @@
 
 const hapiAuthWhitelist = require('./lib/hapi-auth-whitelist')
 const hapiControllersIndex = require('./lib/hapi-controllers-index')
-const hapiControllersPing = require('./lib/hapi-controllers-ping')
 const hapiServer = require('./lib/hapi-server')
 
 const controllers = {
-  index: hapiControllersIndex,
-  ping: hapiControllersPing
+  index: hapiControllersIndex
 }
 const auth = {
   whitelist: hapiAuthWhitelist
