+++ bat-utils/boot-runtime.js
@@ -6,28 +6,22 @@ const _ = require('underscore')
 
 const cache = require('./lib/runtime-cache')
 const currency = require('./lib/runtime-currency')
-const database = require('./lib/runtime-database')
 const kafka = require('./lib/runtime-kafka')
 const newrelic = require('./lib/runtime-newrelic')
 const postgres = require('./lib/runtime-postgres')
 const prometheus = require('./lib/runtime-prometheus')
-const queue = require('./lib/runtime-queue')
 const sentry = require('./lib/runtime-sentry')
-const slack = require('./lib/runtime-slack')
 const wallet = require('./lib/runtime-wallet')
 const wreck = require('./lib/runtime-wreck')
 
 const hash = {
   cache,
   currency,
-  database,
   kafka,
   newrelic,
   postgres,
   prometheus,
-  queue,
   sentry,
-  slack,
   wallet,
   wreck
 }
