+++ bat-utils/lib/runtime-cache.js
@@ -72,6 +72,7 @@ function Cache (config, runtime) {
   } = config
   const { redis } = cache
 
+  console.log('cache config', config)
   if (!redis) {
     return
   }
