+++ docker-compose.yml
@@ -58,7 +58,7 @@ services:
       - "POSTGRES_PASSWORD=password"
   eyeshade-consumer:
     container_name: eyeshade-consumer
-    command: "npm run eyeshade-consumer"
+    command: "yarn run eyeshade-consumer"
     extends:
       file: base.yml
       service: eyeshade
@@ -77,7 +77,7 @@ services:
     container_name: eyeshade-web
     ports:
       - "3002:3002"
-    command: "npm run eyeshade-server"
+    command: "yarn run eyeshade-server"
     extends:
       file: base.yml
       service: eyeshade
@@ -88,93 +88,12 @@ services:
       - "SERVICE=eyeshade"
       - "DYNO=web.1"
     depends_on:
-      - mongo
-      - redis
+      - eyeshade-redis
       - eyeshade-postgres
     networks:
       - eyeshade
     volumes:
       - ./eyeshade:/usr/src/app/eyeshade
-
-  grant-web:
-    container_name: grant-web
-    image: brave/bat-go:grants
-    ports:
-      - "3336:3333"
-    environment:
-      - "DATABASE_URL=postgres://grants:password@grant-postgres/grants?sslmode=disable"
-      - "DATABASE_MIGRATIONS_URL=file:///migrations"
-      - BAT_SETTLEMENT_ADDRESS
-      - CHALLENGE_BYPASS_SERVER=http://challenge-bypass:2416
-      - CHALLENGE_BYPASS_TOKEN
-      - GRANT_SIGNATOR_PUBLIC_KEY
-      - GRANT_WALLET_CARD_ID
-      - GRANT_WALLET_PRIVATE_KEY
-      - GRANT_WALLET_PUBLIC_KEY
-      - "TOKEN_LIST=596fde96-cc80-40e5-b81f-a21c81bb101a"
-      - UPHOLD_ACCESS_TOKEN
-    depends_on:
-      - grant-postgres
-      - challenge-bypass
-      - kafka
-    networks:
-      - eyeshade
-  grant-postgres:
-    container_name: grant-postgres
-    image: postgres:11.9
-    environment:
-      - "POSTGRES_USER=grants"
-      - "POSTGRES_PASSWORD=password"
-    ports:
-      - "3014:5432"
-    networks:
-      - eyeshade
-
-  rewards-dev:
-    container_name: rewards-dev
-    image: brave/rewards-api:latest
-    ports:
-      - "3338:3333"
-    volumes:
-      - ../:/src
-    working_dir: /src
-    networks:
-      - eyeshade
-    environment:
-      - PPROF_ENABLED
-      - ENV=local
-      - ADDR=:3333
-      - "RATIOS_SERVICE=${BAT_RATIOS_URL}"
-      - "RATIOS_TOKEN=${BAT_RATIOS_TOKEN}"
-      - DEFAULT_TIP_CHOICES=1,10,100
-      - DEFAULT_MONTHLY_CHOICES=1,10,100
-      - DEFAULT_AC_CHOICES
-      - DEFAULT_CURRENCY=USD
-      - RATIOS_CACHE_EXPIRY=1m
-      - RATIOS_CACHE_PURGE=2m
-
-  challenge-bypass-postgres:
-    container_name: challenge-bypass-postgres
-    image: postgres:11.9
-    environment:
-      - "POSTGRES_USER=btokens"
-      - "POSTGRES_PASSWORD=password"
-    networks:
-      - eyeshade
-  challenge-bypass:
-    container_name: challenge-bypass
-    image: brave/challenge-bypass
-    restart: always
-    ports:
-      - "2416:2416"
-    environment:
-      - "SENTRY_DSN"
-      - "DATABASE_URL=postgres://btokens:password@challenge-bypass-postgres/btokens?sslmode=disable"
-      - "DATABASE_MIGRATIONS_URL=file:///src/migrations"
-    depends_on:
-      - challenge-bypass-postgres
-    networks:
-      - eyeshade
 
   prometheus:
     image: prom/prometheus:latest
