+++ eyeshade/controllers/referrals.js
@@ -1,5 +1,5 @@
 const boom = require('@hapi/boom')
-const Joi = require('@hapi/joi')
+const Joi = require('joi')
 const underscore = require('underscore')
 const _ = underscore
