+++ eyeshade/lib/queries.js
@@ -1,4 +1,4 @@
-const uuidv5 = require('uuid/v5')
+const { v5: uuidv5 } = require('uuid')
 
 module.exports = {
   allSettlements,
