+++ eyeshade/workers/referrals.js
@@ -5,67 +5,6 @@ const referrals = require('../lib/referrals')
 const queries = require('../lib/queries')
 const countries = require('../lib/countries')
 
-exports.initialize = async (debug, runtime) => {
-  await runtime.queue.create('referral-report')
-}
-
-exports.workers = {
-/* sent by POST /v1/referrals/{transactionId}
-
-    { queue            : 'referral-report'
-    , message          :
-      { transactionId  : '...' }
-    }
-*/
-  'referral-report':
-    async (debug, runtime, payload) => {
-      const referrals = runtime.database.get('referrals', debug)
-      const publishers = runtime.database.get('publishers', debug)
-      const { transactionId } = payload
-      const docs = await referrals.aggregate([
-        {
-          $match: { transactionId }
-        },
-        {
-          $group: {
-            _id: { publisher: '$publisher', owner: '$owner', altcurrency: '$altcurrency' },
-            firstId: { $first: '$_id' },
-            probi: { $sum: '$probi' }
-          }
-        }
-      ])
-
-      const client = await runtime.postgres.connect()
-      try {
-        await client.query('BEGIN')
-        try {
-          for (let i = 0; i < docs.length; i += 1) {
-            const doc = docs[i]
-            if (!doc._id.owner) {
-              const pub = await publishers.findOne({ publisher: doc._id.publisher })
-              if (pub) {
-                if (pub.owner) {
-                  doc._id.owner = pub.owner
-                } else if (pub.authority) {
-                  doc._id.owner = pub.authority
-                }
-              }
-            }
-            await transaction.insertFromReferrals(runtime, client, Object.assign(doc, { transactionId }))
-          }
-        } catch (e) {
-          await client.query('ROLLBACK')
-          runtime.captureException(e, { extra: { report: 'referral-report', transactionId } })
-          throw e
-        }
-        await client.query('COMMIT')
-        runtime.prometheus.getMetric('referral_inserted_counter').inc(docs.length)
-      } finally {
-        client.release()
-      }
-    }
-}
-
 const getTransactionsById = `
 SELECT id
 FROM transactions
