+++ package.json
@@ -5,17 +5,17 @@
   "main": "index.js",
   "scripts": {
     "ava": "ava",
-    "start-eyeshade-server": "NODE_PATH=. DEBUG=*,-mongo:*,mongo:queries SERVICE=eyeshade node --optimize_for_size --max_old_space_size=460 --gc_interval=100 eyeshade/server.js",
+    "start-eyeshade-web": "NODE_PATH=. DEBUG=*,-mongo:*,mongo:queries SERVICE=eyeshade node --optimize_for_size --max_old_space_size=460 --gc_interval=100 eyeshade/server.js",
     "start-eyeshade-consumer": "NODE_PATH=. DEBUG=*,-mongo:*,mongo:queries SERVICE=eyeshade node --optimize_for_size --max_old_space_size=460 --gc_interval=100 eyeshade/consumer.js",
     "start-ledger": "NODE_PATH=. DEBUG=*,-mongo:*,mongo:queries SERVICE=ledger node ledger/server.js",
-    "test": "npx nyc --reporter=html --reporter=text npm run test-all",
+    "test": "npx nyc --reporter=html --reporter=text yarn run test-all",
     "coverage": "npx nyc report --reporter=text-lcov",
-    "test-all": "npm run test-isup && npm run test-unit && npm run test-integration",
+    "test-all": "yarn run test-isup && yarn run test-unit && yarn run test-integration",
     "test-isup": "ava -v -s test/isup.integration.test.js",
-    "docker-test": "docker-compose run --rm -v $(pwd)/test:/usr/src/app/test -v $(pwd)/eyeshade:/usr/src/app/eyeshade -v $(pwd)/ledger:/usr/src/app/ledger -v $(pwd)/bat-utils:/usr/src/app/bat-utils ledger-web npm run test",
+    "docker-test": "docker-compose run --rm -v $(pwd)/test:/usr/src/app/test -v $(pwd)/eyeshade:/usr/src/app/eyeshade -v $(pwd)/ledger:/usr/src/app/ledger -v $(pwd)/bat-utils:/usr/src/app/bat-utils ledger-web yarn run test",
     "lint": "npx standard",
-    "lint-fix": "npm run lint -- --fix",
-    "security": "npm run lint && npm run audit",
+    "lint-fix": "yarn run lint -- --fix",
+    "security": "yarn run lint && yarn run audit",
     "audit": "yarn audit",
     "test-integration": "ava -v -s test/**/*.test.js",
     "test-unit": "ava -v -s bat-utils/**/*.test.js eyeshade/*/*.test.js",
@@ -25,7 +25,7 @@
     "docker-up-dev": "$npm_package_config_dockerfile up $npm_package_config_dockerservices $npm_package_config_dockermonitors",
     "docker-up-dbs": "$npm_package_config_dockerfile up -d --remove-orphans $npm_package_config_dbs",
     "docker-up-detached": "docker-compose up -d $npm_package_config_dockerservices",
-    "docker-migrate-dbs": "npm run docker-up-dbs && npm run docker-migrate && npm run docker-seed-dbs",
+    "docker-migrate-dbs": "yarn run docker-up-dbs && yarn run docker-migrate && yarn run docker-seed-dbs",
     "docker-migrate": "$npm_package_config_dockerfile run --rm -w /usr/src/app/eyeshade eyeshade-web ./bin/migrate-up.sh",
     "docker-seed-dbs": "$npm_package_config_dockerfile run --rm -w /usr/src/app/eyeshade eyeshade-web sh ./bin/seed.sh",
     "docker-migrate-down": "$npm_package_config_dockerfile run --rm -w /usr/src/app/eyeshade eyeshade-web sh ./bin/migrate-down-all.sh",
@@ -34,8 +34,8 @@
   },
   "config": {
     "dockermonitors": "prometheus grafana",
-    "dockerservices": "eyeshade-web eyeshade-consumer grant-web",
-    "dbs": "eyeshade-postgres grant-postgres challenge-bypass-postgres",
+    "dockerservices": "eyeshade-web eyeshade-consumer",
+    "dbs": "eyeshade-postgres",
     "dockerfile": "docker-compose -f docker-compose.yml -f docker-compose.dev.yml"
   },
   "repository": {
@@ -49,13 +49,9 @@
   },
   "homepage": "https://github.com/brave-intl/bat-ledger",
   "dependencies": {
-    "@hapi/bell": "^12.1.1",
     "@hapi/boom": "^9.1.0",
-    "@hapi/cookie": "^11.0.2",
-    "@hapi/cryptiles": "^5.1.0",
     "@hapi/hapi": "^20.0.2",
     "@hapi/inert": "^6.0.3",
-    "@hapi/vision": "^6.0.1",
     "@hapi/wreck": "^17.0.0",
     "@sentry/node": "^5.27.4",
     "@uphold/uphold-sdk-javascript": "^2.4.0",
@@ -79,7 +75,6 @@
     "moment": "^2.29.1",
     "monotonic-timestamp": "0.0.9",
     "netmask": "^1.0.6",
-    "node-slack": "0.0.7",
     "nyc": "^15.1.0",
     "pg": "^8.5.1",
     "prom-client": "^12.0.0",
